+++ axum-extra/src/extract/mod.rs
@@ -1,10 +1,6 @@
 //! Additional extractors.
 
-mod host;
 pub mod rejection;
-
-#[cfg(feature = "optional-path")]
-mod optional_path;
 
 #[cfg(feature = "cached")]
 mod cached;
@@ -26,16 +22,6 @@ mod query;
 
 #[cfg(feature = "multipart")]
 pub mod multipart;
-
-#[cfg(feature = "scheme")]
-mod scheme;
-
-#[allow(deprecated)]
-#[cfg(feature = "optional-path")]
-pub use self::optional_path::OptionalPath;
-
-#[allow(deprecated)]
-pub use self::host::Host;
 
 #[cfg(feature = "cached")]
 pub use self::cached::Cached;
@@ -64,11 +50,6 @@ pub use self::query::{OptionalQueryRejection, Query, QueryRejection};
 
 #[cfg(feature = "multipart")]
 pub use self::multipart::Multipart;
-
-#[allow(deprecated)]
-#[cfg(feature = "scheme")]
-#[doc(no_inline)]
-pub use self::scheme::{Scheme, SchemeMissing};
 
 #[cfg(feature = "json-deserializer")]
 pub use self::json_deserializer::{